requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
authlib==1.4.1
httpx==0.28.1
apscheduler==3.10.4
cryptography==44.0.0
jinja2==3.1.5
python-dotenv==1.0.1
python-multipart==0.0.19
itsdangerous==2.2.0
psycopg2-binary==2.9.10
